消息通知，锚点定位。遗留问题，文章详情

+++ article/views.py
@@ -1,47 +1,88 @@
-from django.contrib.auth.decorators import login_required
-from django.http import HttpResponse
+# 引入redirect重定向模块
 from django.shortcuts import render, redirect, get_object_or_404
-import markdown
-from .forms import ArticlePosrForm
-from .models import ArticlePost, ArticleColumn
+# 引入User模型
 from django.contrib.auth.models import User
+# 引入HttpResponse
+from django.http import HttpResponse
+# 导入数据模型ArticlePost, ArticleColumn
+from .models import ArticlePost, ArticleColumn
+# 引入刚才定义的ArticlePostForm表单类
+from .forms import ArticlePosrForm
+# 引入markdown模块
+import markdown
+# 引入login装饰器
+from django.contrib.auth.decorators import login_required
+# 引入分页模块
 from django.core.paginator import Paginator
+# 引入搜索 Q 对象
 from django.db.models import Q
+# Comment 模型
 from comment.models import Comment
+
 from comment.forms import CommentForm
-# Create your views here.
+
+# 通用类视图
+from django.views import View
+from django.views.generic import ListView, DetailView
+from django.views.generic.edit import CreateView
 
 
+# from my_blog.settings import LOGGING
+# import logging
+
+# logging.config.dictConfig(LOGGING)
+# logger = logging.getLogger('django.request')
+
+
+# 文章列表
 def article_list(request):
+    # 从 url 中提取查询参数
     search = request.GET.get('search')
     order = request.GET.get('order')
     column = request.GET.get('column')
     tag = request.GET.get('tag')
 
+    # 初始化查询集
     article_list = ArticlePost.objects.all()
-    # 用户搜索逻辑
+
+    # 搜索查询集
     if search:
         article_list = article_list.filter(
-                Q(title__icontains=search) |
-                Q(body__icontains=search)
-            ).order_by('-total_views')
+            Q(title__icontains=search) |
+            Q(body__icontains=search)
+        )
     else:
+        # 将 search 参数重置为空
         search = ''
 
-    if column is not  None and column.isdigit():
+    # 栏目查询集
+    if column is not None and column.isdigit():
         article_list = article_list.filter(column=column)
 
+    # 标签查询集
     if tag and tag != 'None':
         article_list = article_list.filter(tags__name__in=[tag])
 
+    # 查询集排序
     if order == 'total_views':
-        article_list = article_list.all().order_by('-total_views')
+        # 按热度排序博文
+        article_list = article_list.order_by('-total_views')
 
+    # 每页显示 1 篇文章
     paginator = Paginator(article_list, 3)
+    # 获取 url 中的页码
     page = request.GET.get('page')
+    # 将导航对象相应的页码内容返回给 articles
     articles = paginator.get_page(page)
-
-    context = {'articles': articles, 'order': order, 'search': search, 'tag':tag}
+    # 需要传递给模板（templates）的对象
+    context = {
+        'articles': articles,
+        'order': order,
+        'search': search,
+        'column': column,
+        'tag': tag,
+    }
+    # render函数：载入模板，并返回context对象
     return render(request, 'article/list.html', context)
 
 
@@ -100,7 +141,9 @@ def article_detail(request, id):
     # 载入模板，并返回context对象
     return render(request, 'article/detail.html', context)
 
+
 # 写文章的视图
+@login_required(login_url='/userprofile/login/')
 def article_create(request):
     # 判断用户是否提交数据
     if request.method == "POST":
@@ -110,12 +153,14 @@ def article_create(request):
         if article_post_form.is_valid():
             # 保存数据，但暂时不提交到数据库中
             new_article = article_post_form.save(commit=False)
-            # 指定目前登录的用户为作者
+            # 指定登录的用户为作者
             new_article.author = User.objects.get(id=request.user.id)
-            # 将新文章保存到数据库中
             if request.POST['column'] != 'none':
+                # 保存文章栏目
                 new_article.column = ArticleColumn.objects.get(id=request.POST['column'])
+            # 将新文章保存到数据库中
             new_article.save()
+            # 保存 tags 的多对多关系
             article_post_form.save_m2m()
             # 完成后返回到文章列表
             return redirect("article:article_list")
@@ -126,29 +171,42 @@ def article_create(request):
     else:
         # 创建表单类实例
         article_post_form = ArticlePosrForm()
-        # 赋值上下文
+        # 文章栏目
         columns = ArticleColumn.objects.all()
-        context = {'article_post_form':article_post_form, 'columns':columns}
+        # 赋值上下文
+        context = {'article_post_form': article_post_form, 'columns': columns}
         # 返回模板
         return render(request, 'article/create.html', context)
 
-# 删文章
+
+# 删除文章，此方式有 csrf 攻击风险
+@login_required(login_url='/userprofile/login/')
 def article_delete(request, id):
     # 根据 id 获取需要删除的文章
     article = ArticlePost.objects.get(id=id)
+    # 过滤非作者的用户
+    if request.user != article.author:
+        return HttpResponse("抱歉，你无权修改这篇文章。")
     # 调用.delete()方法删除文章
     article.delete()
     # 完成删除后返回文章列表
     return redirect("article:article_list")
 
+
+# 安全删除文章
+@login_required(login_url='/userprofile/login/')
 def article_safe_delete(request, id):
-    if request.method =='POST':
+    if request.method == 'POST':
         article = ArticlePost.objects.get(id=id)
+        if request.user != article.author:
+            return HttpResponse("抱歉，你无权修改这篇文章。")
         article.delete()
         return redirect("article:article_list")
     else:
         return HttpResponse("仅允许post请求")
 
+
+# 更新文章
 @login_required(login_url='/userprofile/login/')
 def article_update(request, id):
     """
@@ -211,4 +269,82 @@ def article_update(request, id):
         return render(request, 'article/update.html', context)
 
 
+# 点赞数 +1
+class IncreaseLikesView(View):
+    def post(self, request, *args, **kwargs):
+        article = ArticlePost.objects.get(id=kwargs.get('id'))
+        article.likes += 1
+        article.save()
+        return HttpResponse('success')
 
+
+def article_list_example(request):
+    """
+    与下面的类视图做对比的函数
+    简单的文章列表
+    """
+    if request.method == 'GET':
+        articles = ArticlePost.objects.all()
+        context = {'articles': articles}
+        return render(request, 'article/list.html', context)
+
+
+class ContextMixin:
+    """
+    Mixin
+    """
+
+    def get_context_data(self, **kwargs):
+        # 获取原有的上下文
+        context = super().get_context_data(**kwargs)
+        # 增加新上下文
+        context['order'] = 'total_views'
+        return context
+
+
+class ArticleListView(ContextMixin, ListView):
+    """
+    文章列表类视图
+    """
+    # 查询集的名称
+    context_object_name = 'articles'
+    # 模板
+    template_name = 'article/list.html'
+
+    def get_queryset(self):
+        """
+        查询集
+        """
+        queryset = ArticlePost.objects.filter(title='Python')
+        return queryset
+
+
+class ArticleDetailView(DetailView):
+    """
+    文章详情类视图
+    """
+    queryset = ArticlePost.objects.all()
+    context_object_name = 'article'
+    template_name = 'article/detail.html'
+
+    def get_object(self):
+        """
+        获取需要展示的对象
+        """
+        # 首先调用父类的方法
+        obj = super(ArticleDetailView, self).get_object()
+        # 浏览量 +1
+        obj.total_views += 1
+        obj.save(update_fields=['total_views'])
+        return obj
+
+
+class ArticleCreateView(CreateView):
+    """
+    创建文章的类视图
+    """
+    model = ArticlePost
+    fields = '__all__'
+    # 或者有选择的提交字段，比如：
+    # fields = ['title']
+    template_name = 'article/create_by_class_view.html'
